Renamed MacaroonSerializationFormat to MacaroonsSerializer
The logic that was in MacaroonsSerializer/MacaroonsDeSerializer is now
in MacaroonsSerializerV1. MacaroonsSerializer is now the overall
interface, while MacaroonsDeSerializer has been deleted. The unit tests
of those classes have been merged into MacaroonsSerializerV1Test.
