+++ src/main/java/com/github/nitram509/jmacaroons/MacaroonsBuilder.java
@@ -115,15 +115,15 @@ public class MacaroonsBuilder {
   }
 
   /**
-   * Deserializes a macaroon using the {@link MacaroonSerializationFormat#V1} format.
+   * Deserializes a macaroon using the {@link MacaroonsSerializer#V1} format.
    *
    * @param serializedMacaroon serializedMacaroon
    * @return {@link com.github.nitram509.jmacaroons.Macaroon}
    * @throws com.github.nitram509.jmacaroons.NotDeSerializableException when serialized macaroon is not valid base64, length is to short or contains invalid packet data
-   * @see #deserialize(String, MacaroonSerializationFormat)
+   * @see #deserialize(String, MacaroonsSerializer)
    */
   public static Macaroon deserialize(String serializedMacaroon) throws IllegalArgumentException {
-    return deserialize(serializedMacaroon, MacaroonSerializationFormat.V1);
+    return deserialize(serializedMacaroon, MacaroonsSerializer.V1);
   }
 
   /**
@@ -133,9 +133,9 @@ public class MacaroonsBuilder {
    * @param format the serialization format.
    * @return {@link com.github.nitram509.jmacaroons.Macaroon}
    * @throws com.github.nitram509.jmacaroons.NotDeSerializableException when serialized macaroon is not valid, length is to short or contains invalid packet data
-   * @see #deserialize(String, MacaroonSerializationFormat)
+   * @see #deserialize(String, MacaroonsSerializer)
    */
-  public static Macaroon deserialize(String serializedMacaroon, MacaroonSerializationFormat format) throws IllegalArgumentException {
+  public static Macaroon deserialize(String serializedMacaroon, MacaroonsSerializer format) throws IllegalArgumentException {
     return format.deserialize(serializedMacaroon);
   }
