--- src/main/java/com/github/nitram509/jmacaroons/SerializationFormatV1.java
@@ -1,29 +1,0 @@
-/*
- * Copyright (c) 2019 Neil Madden.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.github.nitram509.jmacaroons;
-
-final class SerializationFormatV1 implements MacaroonSerializationFormat {
-    @Override
-    public String serialize(Macaroon macaroon) {
-        return MacaroonsSerializer.serialize(macaroon);
-    }
-
-    @Override
-    public Macaroon deserialize(String serialized) {
-        return MacaroonsDeSerializer.deserialize(serialized);
-    }
-}
