+++ src/test/java/com/github/nitram509/jmacaroons/MacaroonsSerializerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2014 Martin W. Kirst
+ * Copyright (c) 2019 Neil Madden.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,51 +16,54 @@
 
 package com.github.nitram509.jmacaroons;
 
-import org.testng.annotations.BeforeMethod;
-import org.testng.annotations.Test;
-
 import static org.fest.assertions.Assertions.assertThat;
+
+import org.testng.annotations.DataProvider;
+import org.testng.annotations.Test;
 
 public class MacaroonsSerializerTest {
 
-  private String identifier;
-  private String secret;
-  private String location;
+    // Implements the serialization tests described in https://github.com/rescrv/libmacaroons/blob/master/doc/tests.txt
+    // The v1 examples are double base64-encoded for reasons that are not entirely clear, so I have removed one layer of
+    // base64-encoding from the test data.
 
-  @BeforeMethod
-  public void setUp() throws Exception {
-    location = "http://mybank/";
-    secret = "this is our super secret key; only we should know it";
-    identifier = "we used our secret key";
-  }
+    @DataProvider
+    public Object[][] testCases() {
+        return new Object[][] {
+                {
+                        "MDAyMWxvY2F0aW9uIGh0dHA6Ly9leGFtcGxlLm9yZy8KMDAxNWlkZW50aWZpZXIga2V5aWQKMDAyZnNpZ25hdHVyZSB83ue"
+                                + "SURxbxvUoSFgF3-myTnheKOKpkwH51xHGCeOO9wo",
+                        "AgETaHR0cDovL2V4YW1wbGUub3JnLwIFa2V5aWQAAAYgfN7nklEcW8b1KEhYBd_psk54XijiqZMB-dcRxgnjjvc"
+                },
+                {
+                        "MDAyMWxvY2F0aW9uIGh0dHA6Ly9leGFtcGxlLm9yZy8KMDAxNWlkZW50aWZpZXIga2V5aWQKMDAxZGNpZCBhY2NvdW50ID0"
+                                + "gMzczNTkyODU1OQowMDJmc2lnbmF0dXJlIPVIB_bcbt-Ivw9zBrOCJWKjYlM9v3M5umF2XaS9JZ2HCg",
+                        "AgETaHR0cDovL2V4YW1wbGUub3JnLwIFa2V5aWQAAhRhY2NvdW50ID0gMzczNTkyODU1OQAABiD1SAf23G7fiL8PcwazgiV"
+                                + "io2JTPb9zObphdl2kvSWdhw"
+                },
+                {
+                        "MDAyMWxvY2F0aW9uIGh0dHA6Ly9leGFtcGxlLm9yZy8KMDAxNWlkZW50aWZpZXIga2V5aWQKMDAxZGNpZCBhY2NvdW50ID0"
+                                + "gMzczNTkyODU1OQowMDE1Y2lkIHVzZXIgPSBhbGljZQowMDJmc2lnbmF0dXJlIEvpZ80eoMaya69qSpTumwWxWI"
+                                + "baC6hejEKpPI0OEl78Cg",
+                        "AgETaHR0cDovL2V4YW1wbGUub3JnLwIFa2V5aWQAAhRhY2NvdW50ID0gMzczNTkyODU1OQACDHVzZXIgPSBhbGljZQAABiB"
+                                + "L6WfNHqDGsmuvakqU7psFsViG2guoXoxCqTyNDhJe_A"
+                }
 
-  @Test
-  public void Macaroon_can_be_serialized() {
-    Macaroon m = new MacaroonsBuilder(location, secret, identifier).getMacaroon();
+        };
+    }
 
-    assertThat(MacaroonsSerializer.serialize(m)).isEqualTo("MDAxY2xvY2F0aW9uIGh0dHA6Ly9teWJhbmsvCjAwMjZpZGVudGlmaWVyIHdlIHVzZWQgb3VyIHNlY3JldCBrZXkKMDAyZnNpZ25hdHVyZSDj2eApCFJsTAA5rhURQRXZf91ovyujebNCqvD2F9BVLwo");
-    assertThat(MacaroonsSerializer.serialize(m)).isEqualTo(m.serialize());
-  }
+    @Test(dataProvider = "testCases")
+    public void shouldMatchSerializationTests(String v1, String v2) {
+        // Given
 
-  @Test
-  public void Macaroon_with_caveat_can_be_serialized() {
-    Macaroon m = new MacaroonsBuilder(location, secret, identifier)
-        .add_first_party_caveat("account = 3735928559")
-        .getMacaroon();
+        // When
+        Macaroon v1Macaroon = MacaroonsBuilder.deserialize(v1, MacaroonsSerializer.V1);
+        Macaroon v2Macaroon = MacaroonsBuilder.deserialize(v2, MacaroonsSerializer.V2);
 
-    assertThat(MacaroonsSerializer.serialize(m)).isEqualTo("MDAxY2xvY2F0aW9uIGh0dHA6Ly9teWJhbmsvCjAwMjZpZGVudGlmaWVyIHdlIHVzZWQgb3VyIHNlY3JldCBrZXkKMDAxZGNpZCBhY2NvdW50ID0gMzczNTkyODU1OQowMDJmc2lnbmF0dXJlIB7-R2PykNvODB0IR3Nn4R9O7kVqZJM89mLXl3LbuCEoCg");
-    assertThat(MacaroonsSerializer.serialize(m)).isEqualTo(m.serialize());
-  }
-
-  @Test
-  public void Macaroon_with_3rd_party_caveat_can_be_serialized() {
-    Macaroon m = new MacaroonsBuilder(location, secret, identifier)
-        .add_first_party_caveat("account = 3735928559")
-        .add_third_party_caveat("http://auth.mybank/", "SECRET for 3rd party caveat", identifier)
-        .getMacaroon();
-
-    assertThat(MacaroonsSerializer.serialize(m)).startsWith("MDAxY2xvY2F0aW9uIGh0dHA6Ly9teWJhbmsvCjAwMjZpZGVudGlmaWVyIHdlIHVzZWQgb3VyIHNlY3JldCBrZXkKMDAxZGNpZCBhY2NvdW50ID0gMzczNTkyODU1OQowMDFmY2lkIHdlIHVzZWQgb3VyIHNlY3JldCBrZXkKMDA1MXZpZC");
-    assertThat(MacaroonsSerializer.serialize(m)).isEqualTo(m.serialize());
-  }
+        // Then
+        assertThat(v1Macaroon).isEqualTo(v2Macaroon);
+        assertThat(v2Macaroon.serialize(MacaroonsSerializer.V1)).isEqualTo(v1);
+        assertThat(v1Macaroon.serialize(MacaroonsSerializer.V2)).isEqualTo(v2);
+    }
 
 }
